+++ RamenShopApp/View/ShopSearch/Shop/ReviewingShop/ReviewingSheetType.swift
@@ -1,0 +1,13 @@
+//
+//  ReviewingSheetType.swift
+//  RamenShopApp
+//
+//  Created by Koro Saka on 2021-02-25.
+//  Copyright © 2021 Koro Saka. All rights reserved.
+//
+
+import Foundation
+
+enum ReviewingSheetType {
+    case selectingMedia, utilizingMedia
+}
