+++ RamenShopApp/View/ShopSearch/Shop/ShopInfo/EvaluationLabel.swift
